+++ packages/typespec-go/src/lib.ts
@@ -19,8 +19,12 @@ export interface GoEmitterOptions {
   'single-client'?: boolean;
   'stutter'?: string;
   'fix-const-stuttering'?: boolean;
+  /**
+   * @deprecated Use 'generate-samples' instead
+   */
   'generate-examples'?: boolean;
   'factory-gather-all-params'?: boolean;
+  'generate-samples'?: boolean;
 }
 
 const EmitterOptionsSchema: JSONSchemaType<GoEmitterOptions> = {
@@ -93,6 +97,11 @@ const EmitterOptionsSchema: JSONSchemaType<GoEmitterOptions> = {
       description: 'When true, fix stuttering for `const` types and values. The default is false.',
     },
     'generate-examples': {
+      type: 'boolean',
+      nullable: true,
+      description: 'Deprecated. Use generate-samples instead.',
+    },
+    'generate-samples': {
       type: 'boolean',
       nullable: true,
       description: 'When true, generate example tests. The default is false.',